Standartize endpoints

+++ src/main/java/com/blessedbits/SchoolHub/controllers/SchoolController.java
@@ -8,20 +8,27 @@ import com.blessedbits.SchoolHub.dto.SchoolContactsDto;
 import com.blessedbits.SchoolHub.dto.SchoolInfoDto;
 import com.blessedbits.SchoolHub.dto.UpdateSchoolInfoDto;
 import com.blessedbits.SchoolHub.misc.CloudFolder;
+import com.blessedbits.SchoolHub.misc.RoleBasedAccessUtils;
+import com.blessedbits.SchoolHub.models.News;
 import com.blessedbits.SchoolHub.models.Achievement;
 import com.blessedbits.SchoolHub.models.School;
 import com.blessedbits.SchoolHub.models.SchoolContacts;
 import com.blessedbits.SchoolHub.models.SchoolGallery;
 import com.blessedbits.SchoolHub.models.UserEntity;
-import com.blessedbits.SchoolHub.repositories.SchoolGalleryRepository;
+import com.blessedbits.SchoolHub.projections.dto.ClassDto;
+import com.blessedbits.SchoolHub.projections.dto.CourseDto;
+import com.blessedbits.SchoolHub.projections.dto.SchoolDto;
+import com.blessedbits.SchoolHub.projections.dto.UserDto;
+import com.blessedbits.SchoolHub.projections.mappers.BasicDtoMapper;
+import com.blessedbits.SchoolHub.projections.mappers.SchoolMapper;
 import com.blessedbits.SchoolHub.repositories.SchoolRepository;
 import com.blessedbits.SchoolHub.repositories.UserRepository;
-import com.blessedbits.SchoolHub.services.StorageService;
-import com.blessedbits.SchoolHub.services.UserService;
-import com.blessedbits.SchoolHub.services.SchoolService;
+import com.blessedbits.SchoolHub.repositories.SchoolGalleryRepository;
+import com.blessedbits.SchoolHub.services.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
@@ -46,23 +53,35 @@ public class SchoolController {
     private final UserService userService;
     private final UserRepository userRepository;
     private final SchoolService schoolService;
+    private final ClassService classService;
+    private final CourseService courseService;
     private final SchoolGalleryRepository schoolGalleryRepository;
 
     @Autowired
-    public SchoolController(SchoolService schoolService, SchoolRepository schoolRepository, StorageService storageService, UserService userService, UserRepository userRepository, SchoolGalleryRepository schoolGalleryRepository) {
+    public SchoolController(SchoolService schoolService, SchoolRepository schoolRepository, StorageService storageService, UserService userService, UserRepository userRepository, ClassService classService, CourseService courseService, SchoolGalleryRepository schoolGalleryRepository) {
         this.schoolRepository = schoolRepository;
         this.storageService = storageService;
         this.userService = userService;
         this.userRepository = userRepository;
         this.schoolService = schoolService;
+        this.classService = classService;
+        this.courseService = courseService;
         this.schoolGalleryRepository = schoolGalleryRepository;
     }
 
-    @GetMapping("/")
-    public ResponseEntity<List<School>> getSchools() {
-        return new ResponseEntity<>(schoolRepository.findAll(), HttpStatus.OK);
+    @GetMapping("")
+    public ResponseEntity<List<SchoolDto>> getSchools(
+            @RequestParam(required = false) List<String> include
+    ) {
+        return new ResponseEntity<>(schoolService.getAllAsDto(include), HttpStatus.OK);
     }
 
+    @PostMapping("")
+    public ResponseEntity<String> createSchool(@RequestBody CreateSchoolDto schoolDto) {
+        School school = new School();
+        school.setName(schoolDto.getName());
+        school.setAddress(schoolDto.getAddress());
+      
     @GetMapping("/school")
     public ResponseEntity<?> getSchool(@RequestHeader("Authorization") String authorizationHeader) {
         Integer schoolId = userService.getUserFromHeader(authorizationHeader).getSchool().getId();
@@ -75,33 +94,56 @@ public class SchoolController {
         }
     }
 
-    @PostMapping("/new")
-    public ResponseEntity<?> createSchool(@RequestBody CreateSchoolDto schoolDto) {
+    @GetMapping("/{id}")
+    public ResponseEntity<SchoolDto> getSchool(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getById(id);
+        SchoolDto schoolDto;
+        if (RoleBasedAccessUtils.canAccessSchool(user, school)) {
+            schoolDto = SchoolMapper.INSTANCE.toSchoolDto(school, include);
+        } else {
+            schoolDto = BasicDtoMapper.toSchoolDto(school);
+        }
+        return new ResponseEntity<>(schoolDto, HttpStatus.OK);
+    }
+
+    @PutMapping("/{id}/info")
+    public ResponseEntity<String> updateInfo(
+            @PathVariable Integer id,
+            @RequestBody CreateSchoolDto schoolDto,
+            @AuthenticationPrincipal UserEntity user) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifySchool(user, school)) {
+            return new ResponseEntity<>("You can't modify this school", HttpStatus.FORBIDDEN);
+        }
+        String name = schoolDto.getName();
+        if (name != null && !name.isEmpty()) {
+            school.setName(name);
+        }
+        String address = schoolDto.getAddress();
+        if (address != null && !address.isEmpty()) {
+            school.setAddress(address);
+        }
         try {
-            School school = schoolService.createSchool(schoolDto);
-            return new ResponseEntity<>(school, HttpStatus.CREATED);
+            schoolRepository.save(school);
         } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Unable to update info", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PutMapping("/update-info")
-    public ResponseEntity<String> updateInfo(@RequestBody UpdateSchoolInfoDto schoolDto,
-                                             @RequestHeader("Authorization") String authorizationHeader) {
-        try{
-            School school = userService.getUserFromHeader(authorizationHeader).getSchool();
-            schoolService.updateSchoolInfo(school, schoolDto);
-            return new ResponseEntity<>("Data updated", HttpStatus.OK);
-        }catch(Exception e)
-        {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+    @PutMapping("/{id}/logo")
+    public ResponseEntity<String> updateLogo(
+            @PathVariable Integer id,
+            @RequestParam MultipartFile logo,
+            @AuthenticationPrincipal UserEntity user) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifySchool(user, school)) {
+            return new ResponseEntity<>("You can't modify this school", HttpStatus.FORBIDDEN);
         }
-    }
-
-    @PutMapping("/update-logo")
-    public ResponseEntity<String> updateLogo(@RequestParam MultipartFile logo,
-                                             @RequestHeader("Authorization") String authorizationHeader) {
-        School school = userService.getUserFromHeader(authorizationHeader).getSchool();
         try {
             if (school.getLogo() != null && !school.getLogo().isEmpty()) {
                 storageService.deleteFile(school.getLogo());
@@ -111,29 +153,65 @@ public class SchoolController {
             schoolRepository.save(school);
             return new ResponseEntity<>("Image updated", HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Unable to update logo", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
-    @PostMapping("/add-user")
-    public ResponseEntity<String> addUser(@RequestBody AddSchoolUserDto addSchoolUserDto,
-                                          @RequestHeader("Authorization") String authorizationHeader) {
-        UserEntity user = userService.getUserFromHeader(authorizationHeader);
-        School school;
-        String schoolName = addSchoolUserDto.getSchoolName();
-        if (schoolName == null || schoolName.isEmpty()) {
-            school = user.getSchool();
-        } else {
-            Optional<School> schoolOptional = schoolRepository.findByName(schoolName);
-            if (schoolOptional.isEmpty()) {
-                return new ResponseEntity<>("School with specified name not found",
-                        HttpStatus.NOT_FOUND);
-            }
-            school = schoolOptional.get();
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteSchool(
+            @PathVariable Integer id,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifySchool(user, school)) {
+            return new ResponseEntity<>("You can't modify this school", HttpStatus.FORBIDDEN);
         }
-        user.setSchool(school);
         try {
-            userRepository.save(user);
+            schoolRepository.delete(school);
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Unable to delete school", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+        return new ResponseEntity<>("School deleted", HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}/rating")
+    public ResponseEntity<Map<String, Object>> getRating(
+            @PathVariable Integer id
+    ) {
+        Object[] columns = schoolRepository.findSchoolAverageMarks(id);
+        Map<String, Object> result = new HashMap<>();
+        result.put("schoolName", columns[0]);
+        result.put("averageGrade", columns[1]);
+        return new ResponseEntity<>(result, HttpStatus.OK);
+//        List<Object[]> results = schoolRepository.findSchoolsWithAverageMarks();
+//        return new ResponseEntity<>(
+//                results.stream().map(row -> {
+//                    Map<String, Object> map = new HashMap<>();
+//                    map.put("schoolName", row[0]);
+//                    map.put("averageGrade", row[1]);
+//                    return map;
+//                }).collect(Collectors.toList()), HttpStatus.OK
+//        );
+    }
+
+    @PostMapping("/{id}/users")
+    public ResponseEntity<String> addUser(
+            @PathVariable Integer id,
+            @RequestBody AddSchoolUserDto addSchoolUserDto,
+            @AuthenticationPrincipal UserEntity user) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifySchool(user, school)) {
+            return new ResponseEntity<>("You can't modify this school", HttpStatus.FORBIDDEN);
+        }
+        UserEntity targetUser = userService.getByUsername(addSchoolUserDto.getUsername());
+        if (!RoleBasedAccessUtils.canModifyUser(user, targetUser)) {
+            return new ResponseEntity<>("You can't modify this user", HttpStatus.FORBIDDEN);
+        }
+        targetUser.setSchool(school);
+        try {
+            userRepository.save(targetUser);
             return new ResponseEntity<>("User added", HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>("Couldn't add user to specified school",
@@ -141,17 +219,61 @@ public class SchoolController {
         }
     }
 
-    @GetMapping("/rating")
-    public ResponseEntity<List<Map<String, Object>>> getRating() {
-        List<Object[]> results = schoolRepository.findSchoolsWithAverageMarks();
-        return new ResponseEntity<>(
-                results.stream().map(row -> {
-                    Map<String, Object> map = new HashMap<>();
-                    map.put("schoolName", row[0]);
-                    map.put("averageGrade", row[1]);
-                    return map;
-                }).collect(Collectors.toList()), HttpStatus.OK
-        );
+    @GetMapping("/{id}/users")
+    public ResponseEntity<List<UserDto>> getUsers(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canAccessSchool(user, school)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        List<UserDto> users = userService.mapAllToDto(schoolService.getSchoolUsersLoaded(id, include), include);
+        return new ResponseEntity<>(users, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}/classes")
+    public ResponseEntity<List<ClassDto>> getClasses(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canAccessSchool(user, school)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        List<ClassDto> classes = classService
+                .mapAllToDto(schoolService.getSchoolClassesLoaded(id, include), include);
+        return new ResponseEntity<>(classes, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}/courses")
+    public ResponseEntity<List<CourseDto>> getCourses(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canAccessSchool(user, school)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        List<CourseDto> courses = courseService
+                .mapAllToDto(schoolService.getSchoolCoursesLoaded(id, include), include);
+        return new ResponseEntity<>(courses, HttpStatus.OK);
+    }
+
+    @GetMapping("/{id}/news")
+    public ResponseEntity<List<News>> getNews(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        School school = schoolService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canAccessSchool(user, school)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        return new ResponseEntity<>(schoolService.getSchoolNewsLoaded(id, include), HttpStatus.OK);
     }
 
     @PostMapping("/add-gallery-image")
