+++ src/main/java/com/blessedbits/SchoolHub/controllers/UserController.java
@@ -4,9 +4,13 @@ import com.blessedbits.SchoolHub.dto.UpdateInfoDto;
 import com.blessedbits.SchoolHub.dto.UpdateNameDto;
 import com.blessedbits.SchoolHub.dto.UserProfileDto;
 import com.blessedbits.SchoolHub.misc.CloudFolder;
+import com.blessedbits.SchoolHub.misc.RoleBasedAccessUtils;
 import com.blessedbits.SchoolHub.models.Submission;
 import com.blessedbits.SchoolHub.models.UserEntity;
 import com.blessedbits.SchoolHub.models.VerificationToken;
+import com.blessedbits.SchoolHub.projections.dto.UserDto;
+import com.blessedbits.SchoolHub.projections.mappers.BasicDtoMapper;
+import com.blessedbits.SchoolHub.projections.mappers.UserMapper;
 import com.blessedbits.SchoolHub.repositories.SubmissionRepository;
 import com.blessedbits.SchoolHub.repositories.UserRepository;
 import com.blessedbits.SchoolHub.repositories.VerificationTokenRepository;
@@ -20,6 +24,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -56,12 +61,47 @@ public class UserController {
         this.submissionRepository = submissionRepository;
     }
 
-    @PutMapping("/update-info")
+    @GetMapping("/{id}")
+    public ResponseEntity<UserDto> getUser(
+            @PathVariable Integer id,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        UserEntity loadedUser = userService.getLoadedByIdOrUsername(id, user.getUsername(), include);
+        return new ResponseEntity<>(UserMapper.INSTANCE.toUserDto(loadedUser, include), HttpStatus.OK);
+    }
+
+    @GetMapping("")
+    public ResponseEntity<UserDto> getUser(
+            @RequestParam(required = false) String username,
+            @RequestParam(required = false) String email,
+            @RequestParam(required = false) List<String> include,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        if (username != null) {
+            UserEntity loadedUser = userService.getLoadedByUsername(username, include);
+            if (RoleBasedAccessUtils.canAccessUser(user, loadedUser)) {
+                return new ResponseEntity<>(UserMapper.INSTANCE.toUserDto(loadedUser, include), HttpStatus.OK);
+            }
+            return new ResponseEntity<>(BasicDtoMapper.toUserDto(loadedUser), HttpStatus.OK);
+        }
+        if (email != null) {
+            UserEntity loadedUser = userService.getLoadedByEmail(email, include);
+            if (RoleBasedAccessUtils.canAccessUser(user, loadedUser)) {
+                return new ResponseEntity<>(UserMapper.INSTANCE.toUserDto(loadedUser, include), HttpStatus.OK);
+            }
+            return new ResponseEntity<>(BasicDtoMapper.toUserDto(loadedUser), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
+
+    @PutMapping("/{id}/info")
     public ResponseEntity<String> updateUserInfo(
-            @RequestHeader("Authorization") String authorizationHeader,
-            @RequestBody @Valid UpdateInfoDto updateInfoDto)
-    {
-        UserEntity user = userService.getUserFromHeader(authorizationHeader);
+            @PathVariable Integer id,
+            @RequestBody @Valid UpdateInfoDto updateInfoDto,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        UserEntity targetUser = userService.getByIdOrUser(id, user);
         String email = updateInfoDto.getEmail();
         String username = updateInfoDto.getUsername();
         if(username != null && !username.isEmpty())
@@ -70,7 +110,7 @@ public class UserController {
             {
                 return new ResponseEntity<>("Username already taken!", HttpStatus.BAD_REQUEST);
             }
-            user.setUsername(username);
+            targetUser.setUsername(username);
         }
         if(email != null && !email.isEmpty())
         {
@@ -78,38 +118,88 @@ public class UserController {
             {
                 return new ResponseEntity<>("Email already taken!", HttpStatus.BAD_REQUEST);
             }
-            user.setEmail(email);
+            targetUser.setEmail(email);
             String token = UUID.randomUUID().toString();
             VerificationToken verificationToken = new VerificationToken();
             verificationToken.setToken(token);
-            verificationToken.setUser(user);
+            verificationToken.setUser(targetUser);
             verificationToken.setExpiryDate(LocalDateTime.now().plusMinutes(15));
             verificationTokenRepository.save(verificationToken);
             try{
-                emailService.sendEmail(email,"Please verify your email", emailService.buildConfirmEmail(user.getUsername(), token));
+                emailService.sendEmail(
+                        email,"Please verify your email",
+                        emailService.buildConfirmEmail(targetUser.getUsername(), token));
             } catch (Exception e) {
-                return new ResponseEntity<>(("Failed to send verification token\n" + e), HttpStatus.INTERNAL_SERVER_ERROR);
+                System.out.println(e.getMessage());
+                return new ResponseEntity<>("Failed to send verification token\n",
+                        HttpStatus.INTERNAL_SERVER_ERROR);
             }
-            return new ResponseEntity<>("User info was updated successfully! Please check your email for verification.", HttpStatus.CREATED);
+            return new ResponseEntity<>(
+                    "User info was updated successfully! Please check your email for verification.",
+                    HttpStatus.CREATED);
         }
-        userRepository.save(user);
+        userRepository.save(targetUser);
         return new ResponseEntity<>("User info was updated successfully!", HttpStatus.CREATED);
     }
 
-    @PutMapping("/update-profile-image")
-    public ResponseEntity<String> updateProfileImage(@RequestParam MultipartFile profileImage,
-                                                     @RequestHeader("Authorization") String authorizationHeader) {
-        UserEntity user = userService.getUserFromHeader(authorizationHeader);
+    @PutMapping("/{id}/image")
+    public ResponseEntity<String> updateProfileImage(
+            @PathVariable Integer id,
+            @RequestParam MultipartFile profileImage,
+            @AuthenticationPrincipal UserEntity user) {
+        UserEntity targetUser = userService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifyUser(user, targetUser)) {
+            return new ResponseEntity<>("You can't modify this user", HttpStatus.FORBIDDEN);
+        }
         try {
             if (user.getProfileImage() != null && !user.getProfileImage().isEmpty()) {
                 storageService.deleteFile(user.getProfileImage());
             }
             String url = storageService.uploadFile(profileImage, CloudFolder.PROFILE_IMAGES);
-            user.setProfileImage(url);
-            userRepository.save(user);
+            targetUser.setProfileImage(url);
+            userRepository.save(targetUser);
             return new ResponseEntity<>("Image updated", HttpStatus.OK);
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @PutMapping("/{id}/duty")
+    public ResponseEntity<String> setUsersDuty(
+            @PathVariable Integer id,
+            @RequestBody String duty,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        UserEntity targetUser = userService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifyUser(user,  targetUser)) {
+            return new ResponseEntity<>("You can't modify this user", HttpStatus.FORBIDDEN);
+        }
+        targetUser.setDuty(duty);
+        try {
+            userRepository.save(targetUser);
+            return new ResponseEntity<>("User's duty was successfully updated.", HttpStatus.OK);
+        }
+        catch(Exception e) {
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Unable to update user duty.", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<String> deleteUser(
+            @PathVariable Integer id,
+            @AuthenticationPrincipal UserEntity user
+    ) {
+        UserEntity targetUser = userService.getByIdOrUser(id, user);
+        if (!RoleBasedAccessUtils.canModifyUser(user, targetUser)) {
+            return new ResponseEntity<>("You can't modify this user", HttpStatus.FORBIDDEN);
+        }
+        try {
+            userRepository.delete(targetUser);
+            return new ResponseEntity<>("User deleted", HttpStatus.OK);
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            return new ResponseEntity<>("Unable to delete user", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -135,41 +225,20 @@ public class UserController {
         }
     }
 
-    @GetMapping("/grades")
+    @GetMapping("/{id}/grades")
     public ResponseEntity<List<Submission>> getGrades(
             @RequestParam("startDate") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate startDate,
             @RequestParam("endDate") @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate endDate,
-            @RequestHeader("Authorization") String authorizationHeader) {
-        UserEntity user = userService.getUserFromHeader(authorizationHeader);
+            @AuthenticationPrincipal UserEntity user) {
+        UserEntity targetUser = userService.getByIdOrUser(user.getId(), user);
         try {
             List<Submission> submissions = submissionRepository.findSubmissionsByStudentIdAndDateRange(
-                    user.getId(), startDate.atStartOfDay(), endDate.atStartOfDay()
+                    targetUser.getId(), startDate.atStartOfDay(), endDate.atStartOfDay()
             );
             return new ResponseEntity<>(submissions, HttpStatus.OK);
         } catch (Exception e) {
             System.out.println(e.getMessage());
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    @PutMapping("/set-duty/{id}")
-    public ResponseEntity<String> setUsersDuty(@PathVariable Integer id, @RequestBody String duty) 
-    {
-        Optional<UserEntity> userOpt = userRepository.findById(id);
-        if(userOpt.isEmpty())
-        {
-            return new ResponseEntity<>("Error: User is not found by provided ID.", HttpStatus.NOT_FOUND);
-        }
-        UserEntity user = userOpt.get();
-        user.setDuty(duty);
-        try 
-        {
-            userRepository.save(user);
-            return new ResponseEntity<>("User's duty was successfully updated.", HttpStatus.OK);
-        }
-        catch(Exception e)
-        {
-            return new ResponseEntity<>(("Error: User's duty was not updated." + e), HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
